move enums, structs to Types.h
