+++ firmware/ventilator-controller-stm32/Core/Src/Pufferfish/Driver/Serial/Nonin/Sensor.cpp
@@ -30,7 +30,7 @@ InitializableState Sensor::output(float &spo2, float &hr) {
     return InitializableState::failed;
   }
 
-  if (device_.output(measurements_) == Device::PacketStatus::available) {
+  if (device_.output(measurements_) == PacketStatus::available) {
     if (measurements_.e_spo2_d == value_unavailable) {
       spo2 = NAN;
     } else {
